# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_v1_upload_api_v1_upload_post": {
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_v1_upload_api_v1_upload_post",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Body of every non-2xx ``/api/v1`` response.\n\n``code`` is a stable machine-readable identifier the client branches on;\n``detail`` is the human-readable message it may show as-is.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ProposedCut": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "enum": [
                  "high",
                  "medium",
                  "low"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "confidence",
          "notes"
        ],
        "title": "ProposedCut",
        "type": "object"
      },
...
      "SlicingConfirm": {
        "description": "Split after each listed page (1 \u2264 cut < page_count).",
        "properties": {
          "cuts": {
            "items": {
              "type": "integer"
            },
            "title": "Cuts",
            "type": "array"
          }
        },
        "title": "SlicingConfirm",
        "type": "object"
      },
      "SlicingConfirmed": {
        "properties": {
          "document_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Document Ids",
            "type": "array"
          }
        },
        "required": [
          "document_ids"
        ],
        "title": "SlicingConfirmed",
        "type": "object"
      },
      "SlicingPage": {
        "properties": {
          "has_thumbnail": {
            "title": "Has Thumbnail",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "text_head": {
            "title": "Text Head",
            "type": "string"
          },
          "text_tail": {
            "title": "Text Tail",
            "type": "string"
          }
        },
        "required": [
          "page",
          "text_head",
          "text_tail",
          "has_thumbnail"
        ],
        "title": "SlicingPage",
        "type": "object"
      },
...
      "SlicingView": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/SlicingPage"
            },
            "title": "Pages",
            "type": "array"
          },
          "proposed_cuts": {
            "items": {
              "$ref": "#/components/schemas/ProposedCut"
            },
            "title": "Proposed Cuts",
            "type": "array"
          },
          "status": {
            "enum": [
              "preparing",
              "ready",
              "failed",
              "done"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "required": [
          "batch_id",
          "subject",
          "status",
          "page_count",
          "pages",
          "proposed_cuts",
          "error"
        ],
        "title": "SlicingView",
        "type": "object"
      },
...
      "UploadResponse": {
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "queued": {
            "title": "Queued",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/UploadResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results",
          "queued",
          "failed",
          "batch_id"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
      "UploadResult": {
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "enum": [
              "queued",
              "duplicate",
              "error"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "status"
        ],
        "title": "UploadResult",
        "type": "object"
      },
      "UploadTarget": {
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "case_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Title"
          },
          "parent_options": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Parent Options",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "case_title",
          "parent_options"
        ],
        "title": "UploadTarget",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sanctuary API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/slicing/{batch_id}": {
      "get": {
        "operationId": "v1_slicing_api_v1_slicing__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlicingView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Slicing",
        "tags": [
          "slicing"
        ]
      }
    },
    "/api/v1/slicing/{batch_id}/confirm": {
      "post": {
        "operationId": "v1_confirm_slicing_api_v1_slicing__batch_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlicingConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlicingConfirmed"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Confirm Slicing",
        "tags": [
          "slicing"
        ]
      }
    },
    "/api/v1/slicing/{batch_id}/retry": {
      "post": {
        "description": "Re-run the slice proposal for a batch whose preparation failed.",
        "operationId": "v1_retry_api_v1_slicing__batch_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlicingView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Retry",
        "tags": [
          "slicing"
        ]
      }
    },
    "/api/v1/slicing/{batch_id}/thumb/{page}": {
      "get": {
        "operationId": "v1_thumbnail_api_v1_slicing__batch_id__thumb__page__get",
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "required": true,
            "schema": {
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Thumbnail",
        "tags": [
          "slicing"
        ]
      }
    },
...
    "/api/v1/upload": {
      "post": {
        "description": "Ingest files: non-.eml files share one manual batch; each .eml is its own batch.",
        "operationId": "v1_upload_api_v1_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_v1_upload_api_v1_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Upload",
        "tags": [
          "upload"
        ]
      }
    },
    "/api/v1/upload/target": {
      "get": {
        "description": "What the upload modal needs when opened for a case: title and parent docs.",
        "operationId": "v1_upload_target_api_v1_upload_target_get",
        "parameters": [
          {
            "in": "query",
            "name": "case_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTarget"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Upload Target",
        "tags": [
          "upload"
        ]
      }
    },
...
  }
}